# Truncate discarded packets in last batch