@bw-solana 2023-03-15
Not seeing any material performance benefits in sigverify from these changes. Going to close this PR for now